Add Number of Ways to Split Array

+++ Number_of_Ways_to_Split_Array.cpp
@@ -2,18 +2,18 @@ class Solution {
 public:
     int waysToSplitArray(vector<int>& nums) {
         int n = nums.size();
-        vector<long long> prefixSum(n); 
+        vector<int> prefixSum(n);
 
         prefixSum[0] = nums[0];
         for (int i = 1; i < n; i++) {
             prefixSum[i] = prefixSum[i - 1] + nums[i];
         }
 
-        long long totalSum = prefixSum[n - 1];
+        int totalSum = prefixSum[n - 1];
         int count = 0;
         for (int i = 0; i < n - 1; i++) {
-            long long first = prefixSum[i];
-            long long last = totalSum - first;
+            int first = prefixSum[i];
+            int last = totalSum - first;
 
             if (first >= last) {
                 count++;
